[Refactor] main.js에서 router 분리
