+++ src/main.js
@@ -1,8 +1,5 @@
 import userStore from "./store/userStore";
-import Home from "./pages/Home";
-import Profile from "./pages/Profile";
-import NotFound from "./pages/NotFound";
-import Login from "./pages/Login";
+import { navigate, render } from "./router";
 
 const loginAction = (e) => {
   e.preventDefault();
@@ -27,40 +24,6 @@ const profileUpdateAction = () => {
   const bio = document.getElementById("bio").value;
   userStore.setUser({ ...user, username, email, bio });
   navigate("/profile");
-};
-
-const routes = {
-  "/": () => Home(),
-  "/login": () => {
-    if (userStore.isLoggedIn()) {
-      navigate("/");
-      return Home();
-    }
-    return Login();
-  },
-  "/profile": () => {
-    if (!userStore.isLoggedIn()) {
-      navigate("/login");
-      return Login();
-    }
-    return Profile();
-  },
-};
-
-const navigate = (path) => {
-  window.history.pushState({}, "", path);
-  render(); // TODO: 필요한지 확인
-};
-
-const App = () => {
-  const path = window.location.pathname;
-  const PageComponent = routes[path] || (() => NotFound());
-  return PageComponent();
-};
-
-const render = () => {
-  const root = document.getElementById("root");
-  root.innerHTML = App();
 };
 
 window.addEventListener("popstate", render);
